requirements.txt
flask
gunicorn==23.0.0
python-dotenv
pyjwt
requests
Flask-SQLAlchemy
gunicorn